Begin code reorganization
All the algorithms will stay the same for now, but the tasks
are being redelegated to new classes and structures. The goal
of this task is to make the code easier to read and follow,
more modular (so different algorithms can accomplish the same tasks),
and minimize the ram it takes to construct a 3D object.

The planned changes can be found at https://docs.google.com/document/d/1P15Q9y-lwCI5fmh5z0lqURAq97g7qONUgkK-tXSJJDk/edit#
TrackablePoint has been removed, and its algorithms have been placed in Feature.

TODO replace the View class per the note in the document above